pyproject.toml
[tool.poetry]
name = "fastapi-async-realworld"
version = "0.1.0"
description = "Realworld app API implementation using FastAPI, async SQL database"
authors = ["congdh <congdh2812@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
databases = "^0.3.2"
psycopg2-binary = "^2.8.6"
asyncpg = "^0.21.0"
pydantic = {version = "^1.6.1", extras = ["email"]}
python-jose = {version = "^3.2.0", extras = ["cryptography"]}
passlib = "^1.7.2"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
pytest-asyncio = "^0.14.0"
httpx = "^0.14.3"
black = {version = "^20.8b1", allow-prereleases = true}
isort = "^5.5.2"
autoflake = "^1.4"
flake8 = "^3.8.3"
mypy = "^0.782"
sqlalchemy-stubs = "^0.3"
pre-commit = "^2.7.1"
bandit = "^1.6.2"
safety = "^1.9.0"
Faker = "^4.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
